usePlanにメソッドを移動
