+++ src/App.js
@@ -9,114 +9,10 @@ import Square from "./Square";
 import Plan from "./Plan";
 import {usePlan} from "./hooks/usePlan";
 
-// class App extends React.Component {
-//   // get axios() {
-//   //   const axiosBase = require('axios');
-//   //   return axiosBase.create({
-//   //     baseURL: process.env.REACT_APP_DEV_API_URL,
-//   //     headers: {
-//   //       'Content-Type': 'application/json',
-//   //       'Accept': 'application/json',
-//   //       'X-Requested-With': 'XMLHttpRequest',
-//   //     },
-//   //     responseType: 'json'
-//   //   });
-//   // }
-//
-//
-//   render() {
-//     // TODO:バックエンドの実装
-//     // this.axios.get('/tasks')
-//     //   .then(results => {
-//     //     console.log(results);
-//     //   })
-//     //   .catch(data => {
-//     //     console.log(data);
-//     //   });
-//     return (
-//       <Schedule />
-//     )
-//   }
-// }
-
-const SQUARE_HEIGHT = 80
-const UNIT_NUM_IN_SQUARE = 4
-const UNIT_HEIGHT = SQUARE_HEIGHT / UNIT_NUM_IN_SQUARE
-const UNIT_MINUTES = 60 / UNIT_NUM_IN_SQUARE
-
 
 export default function App(){
   const [text, setText] = useState(localStorage.getItem("text") ?? '')
   const [planKey, setPlanKey] = useState(localStorage.getItem("planKey") ?? 0)
-  const {plan, setPlan} = usePlan()
-
-  const UNSAFE_componentWillMount = () => {
-    formatText(plan);
-  }
-
-  // ヒント：フック間で情報を受け渡す
-  // https://ja.reactjs.org/docs/hooks-custom.html#tip-pass-information-between-hooks
-  const createPlan = (startHour, endHour) => {
-    const updatedPlan = plan.slice(0)
-    updatedPlan.push(
-      {
-        key: planKey.toString(),
-        description: '',
-        startHour:  startHour,
-        endHour:  endHour,
-        startMinute:  0,
-        endMinute:  0,
-        minutes: 60,
-        hours: 1,
-      }
-    )
-    setPlan(updatedPlan)
-    setPlanKey(Number(planKey) + 1)
-    const text = formatText(updatedPlan)
-    updateLocalStorage(text, updatedPlan, Number(planKey) + 1);
-  }
-
-  const updatePlan =
-    (planKey, startHour, endHour, startMinute, endMinute, minutes, hours, description) => {
-    const updatedPlan = plan.slice(0).map(p => {
-      if(p.key === planKey){
-        p.startHour = startHour
-        p.endHour = endHour
-        p.startMinute = startMinute
-        p.endMinute = endMinute
-        p.minutes = minutes
-        p.hours = hours
-        if(description !== ''){
-          plan.description = description
-        }
-      }
-      return p
-    })
-    setPlan(updatedPlan)
-    const text = formatText(plan)
-    updateLocalStorage(text, plan, planKey);
-  }
-
-  const deletePlan = (e) => {
-    let planKey = e.target.getAttribute('plankey')
-    const updatedPlan = plan.slice(0).filter((p) => {
-      return Number(p.key) !== Number(planKey)
-    })
-    setPlan(updatedPlan)
-    const text = formatText(updatedPlan)
-    updateLocalStorage(text, updatedPlan, planKey);
-  }
-
-  const updateLocalStorage = (text, plan, planKey) => {
-    localStorage.setItem("text", text)
-    localStorage.setItem("plan", JSON.stringify(plan))
-    localStorage.setItem("planKey", planKey)
-  }
-
-  const changeText = (e) => {
-    setText(e.target.value)
-    updateLocalStorage(e.target.value, plan, planKey);
-  }
 
   const formatText = (plan) => {
     // TOOD:開始時間が早い順にソートする
@@ -139,6 +35,35 @@ export default function App(){
     }).join("\n")
     setText(text)
     return text
+  }
+
+  const updateLocalStorage = (text, plan, planKey) => {
+    localStorage.setItem("text", text)
+    localStorage.setItem("plan", JSON.stringify(plan))
+    localStorage.setItem("planKey", planKey)
+  }
+
+  // ヒント：フック間で情報を受け渡す
+  // https://ja.reactjs.org/docs/hooks-custom.html#tip-pass-information-between-hooks
+  const {
+    plan,
+    setPlan,
+    createPlan,
+    updatePlan,
+    deletePlan,
+    calculatePlanTime
+  } = usePlan(
+    text,
+    setText,
+    planKey,
+    setPlanKey,
+    formatText,
+    updateLocalStorage
+  )
+
+  const changeText = (e) => {
+    setText(e.target.value)
+    updateLocalStorage(e.target.value, plan, planKey);
   }
 
   const updateZindex = (planKey) => {
@@ -179,36 +104,6 @@ export default function App(){
     }
   }
 
-  const calculatePlanTime = (height, postition) => {
-    const unit = height / UNIT_HEIGHT
-    // const step = unit % UNIT_NUM_IN_SQUARE
-    const minutes = (unit * UNIT_HEIGHT) / SQUARE_HEIGHT * 60
-    const startHour = Math.floor(Math.round((postition / SQUARE_HEIGHT)* 10) / 10)
-    let startMinute = (Math.round((postition % SQUARE_HEIGHT) / UNIT_HEIGHT ) * UNIT_MINUTES) % 60
-    const endHour = startHour + Math.floor((startMinute + minutes) / 60)
-    let endMinute = (startMinute + minutes) % 60
-    // for debug
-    // console.log('-----------------------------')
-    // console.log('SQUARE_HEIGHT : ' + SQUARE_HEIGHT)
-    // console.log('UNIT_HEIGHT : ' + UNIT_HEIGHT)
-    // console.log('height : ' + height)
-    // console.log('postition : ' + postition)
-    // console.log('unit : ' + unit)
-    // console.log('minutes : ' + minutes)
-    // console.log('startHour : ' + startHour)
-    // console.log('endHour : ' + endHour)
-    // console.log('startMinute : ' + startMinute)
-    // console.log('endMinute : ' + endMinute)
-    // console.log('step : ' + step)
-    return {
-      startHour: startHour,
-      endHour: endHour,
-      startMinute: startMinute,
-      endMinute: endMinute,
-      minutes: minutes,
-      hours: minutes / 60,
-    }
-  }
 
   const saveDescription = (e) => {
     let description = e.target.value
